rename feature to module
